Image storage and retrieval.
Version:        0.0.1
Author(s):      Anthony Pizzimenti, Cale Bierman
Date:           February 25, 2017

BaseItemService uses the store() method on the BaseImageService to
store images.

The way this works is that, when an Item is provided to either the
create() or update() methods, the File objects in the Item.images array
are removed, uploaded to the database, and then we set the Item.images
array equal to the array of URLs returned.

When updating an Item, we can simply append a new File object to the
existing Item.images array, and the update() method will find the File
objects, upload them, flatten and remove falsy values from the array,
and save the Item in the db.