Add support for storing user preferences.
Version:        0.0.1
Author(s):      Anthony Pizzimenti, Cale Bierman
Date:           February 25, 2018

When creating a new User, submit preferences object to the .create
method, and the User's preferences will be saved.

        BaseUserService.create(<email>, <password>, <preferences>)...

These are public and editable, so they can be updated by changing them
and then submitting the User object to the .update(<user>) method.